Change bootstrap logic
