+++ pkg/reconciler/instances/istio/action_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/kyma-incubator/reconciler/pkg/reconciler"
 	"github.com/kyma-incubator/reconciler/pkg/reconciler/kubernetes"
+	"go.uber.org/zap"
 
 	"github.com/kyma-incubator/reconciler/pkg/reconciler/chart"
 	actionsmocks "github.com/kyma-incubator/reconciler/pkg/reconciler/instances/istio/actions/mocks"
@@ -58,13 +59,13 @@ metadata:
 `
 )
 
-var _ = func() interface{} {
-	//This code only executes during test run
-	_skipBootstrap = true
-	return nil
-}()
-
 func Test_ReconcileAction_Run(t *testing.T) {
+
+	performerCreatorFn := func(p actions.IstioPerformer) bootstrapIstioPerformer {
+		return func(logger *zap.SugaredLogger) (actions.IstioPerformer, error) {
+			return p, nil
+		}
+	}
 
 	t.Run("should not perform any istio action when provider returned an error ", func(t *testing.T) {
 		// given
@@ -74,7 +75,7 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		kubeClient := newFakeKubeClient()
 		actionContext := newFakeServiceContext(&factory, &provider, kubeClient)
 		performer := actionsmocks.IstioPerformer{}
-		action := ReconcileAction{performer: &performer}
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -100,7 +101,8 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		actionContext := newFakeServiceContext(&factory, &provider, kubeClient)
 		performer := actionsmocks.IstioPerformer{}
 		performer.On("Version", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("*zap.SugaredLogger")).Return(actions.IstioStatus{}, errors.New("Version error"))
-		action := ReconcileAction{performer: &performer}
+
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -134,7 +136,8 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		performer.On("Version", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("*zap.SugaredLogger")).Return(noIstioOnTheCluster, nil)
 		performer.On("Install", mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("*zap.SugaredLogger")).
 			Return(errors.New("Perfomer Install error"))
-		action := ReconcileAction{performer: &performer}
+
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -168,7 +171,8 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		performer.On("Version", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(noIstioOnTheCluster, nil)
 		performer.On("Install", mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(nil)
 		performer.On("PatchMutatingWebhook", actionContext.KubeClient, actionContext.Logger).Return(errors.New("Performer Patch error"))
-		action := ReconcileAction{performer: &performer}
+
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -201,7 +205,8 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		performer.On("Version", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(noIstioOnTheCluster, nil)
 		performer.On("Install", mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(nil)
 		performer.On("PatchMutatingWebhook", actionContext.KubeClient, actionContext.Logger).Return(nil)
-		action := ReconcileAction{performer: &performer}
+
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -237,7 +242,8 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		performer.On("Version", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(tooLowClientVersion, nil)
 		performer.On("Install", mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(nil)
 		performer.On("PatchMutatingWebhook", actionContext.KubeClient, actionContext.Logger).Return(nil)
-		action := ReconcileAction{performer: &performer}
+
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -273,7 +279,8 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		performer.On("Version", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(tooHighPilotAndDataPlaneVersion, nil)
 		performer.On("Install", mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(nil)
 		performer.On("PatchMutatingWebhook", actionContext.KubeClient, actionContext.Logger).Return(nil)
-		action := ReconcileAction{performer: &performer}
+
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -309,7 +316,8 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		performer.On("Version", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(tooLowPilotAndDataPlaneVersion, nil)
 		performer.On("Install", mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(nil)
 		performer.On("PatchMutatingWebhook", actionContext.KubeClient, actionContext.Logger).Return(nil)
-		action := ReconcileAction{performer: &performer}
+
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -347,7 +355,8 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		performer.On("PatchMutatingWebhook", actionContext.KubeClient, actionContext.Logger).Return(nil)
 		performer.On("Update", mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(nil)
 		performer.On("ResetProxy", mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(errors.New("Proxy reset error"))
-		action := ReconcileAction{performer: &performer}
+
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -386,7 +395,8 @@ func Test_ReconcileAction_Run(t *testing.T) {
 		performer.On("PatchMutatingWebhook", actionContext.KubeClient, actionContext.Logger).Return(nil)
 		performer.On("Update", mock.AnythingOfType("string"), mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(nil)
 		performer.On("ResetProxy", mock.AnythingOfType("string"), mock.AnythingOfType("string"), actionContext.Logger).Return(nil)
-		action := ReconcileAction{performer: &performer}
+
+		action := ReconcileAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -430,7 +440,13 @@ func newFakeKubeClient() *k8smocks.Client {
 
 	return mockClient
 }
+
 func Test_UninstallAction(t *testing.T) {
+	performerCreatorFn := func(p actions.IstioPerformer) bootstrapIstioPerformer {
+		return func(logger *zap.SugaredLogger) (actions.IstioPerformer, error) {
+			return p, nil
+		}
+	}
 
 	noIstioOnTheCluster := actions.IstioStatus{
 		ClientVersion:    "1.0",
@@ -458,7 +474,7 @@ func Test_UninstallAction(t *testing.T) {
 		performer.On("Uninstall", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType("*zap.SugaredLogger")).Return(nil)
 		provider.On("RenderManifest", mock.AnythingOfType("*chart.Component")).Return(&chart.Manifest{}, nil)
 
-		action := UninstallAction{performer: &performer}
+		action := UninstallAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -480,7 +496,8 @@ func Test_UninstallAction(t *testing.T) {
 		performer.On("Version", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType(
 			"string"), mock.AnythingOfType("string"), mock.AnythingOfType("*zap.SugaredLogger")).
 			Return(noIstioOnTheCluster, nil)
-		action := UninstallAction{performer: &performer}
+
+		action := UninstallAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
@@ -501,7 +518,8 @@ func Test_UninstallAction(t *testing.T) {
 		performer.On("Version", mock.Anything, mock.AnythingOfType("string"), mock.AnythingOfType(
 			"string"), mock.AnythingOfType("string"), mock.AnythingOfType("*zap.SugaredLogger")).
 			Return(noIstioOnTheCluster, errors.New("error in detecting istio version"))
-		action := UninstallAction{performer: &performer}
+
+		action := UninstallAction{performerCreatorFn(&performer)}
 
 		// when
 		err := action.Run(actionContext)
